Create 5_Healthcare_in_Different_States.py
We’re going to focus on only the ways in which hospitals charge patients for chest pain. But you could do this for any of the diagnoses in our dataset.
Can you find the official name of the diagnosis related to chest pain?
Let’s grab only the rows in the dataset that are about chest pain.
We now want to separate the dataset by state. How to get every chest pain diagnosis in Alabama?
We’re almost there! We now have all of the hospitals in Alabama that have a diagnosed chest pain. We now want to find the average cost of those diagnoses.
Nice work! We’ve made a boxplot for one state. But how does that state compare to the others? Let’s make a boxplot for every state!
Nice work! You should now see 50 boxplots. If the graph is too small to see, you can expand it by clicking on the arrows in the top right corner of the right panel. You can also take a look at this image to see what your final results should look like.
What information have you learned by looking at these boxplots side by side? What state has the largest spread? What state has the largest median? Which states have the most outliers?